refactor: WinningNumberInput 클래스 분리
